Added a README.
Went through the construction of the read me for GIThub, Unsure of some of the mark down but it feels right.

Added a line to the gitignore to ignore the TestImages folder for while creating the ReadMe I found I would need it. Same as the TECHNICALDOC.md, more so to make sure of the markdown link and confirm its integration into GITHUD.